refactor(FileGraphStorageProvider): Extract path and edge index logic to reduce size of FileGraphStorageProvider.cs
