+++ docs/SessionSummary_2025-04-23.md
@@ -1,0 +1,53 @@
+# Session Summary (2025-04-23)
+
+## Goal
+Initial goal was to gain project understanding and then address TODOs. The focus shifted to refactoring long files.
+
+## Work Completed
+
+1.  **Project Familiarization:**
+    *   Read core project documentation files:
+        *   `CONTRIBUTING.md`
+        *   `issues.md`
+        *   `LICENSE`
+        *   `README.md`
+        *   `docs/AI_Assistant_Guidelines.md`
+        *   `docs/DisambiguationStrategies.md`
+        *   `docs/EfficientGraphIO.md`
+        *   `docs/KnowledgeRepresentationV3.md`
+        *   `ReasoningEngine/Core/ProbabilityDistributions.md`
+        *   `ReasoningEngine/GraphFileHandling/FileManagement.md`
+        *   `docs/SessionSummary_2025-04-15.md`
+    *   Read `docs/git-ignore/LocalNeo4jDatabaseSetup.md` upon request.
+
+2.  **Status Check & TODO Review:**
+    *   Checked `git status` (working tree was clean).
+    *   Reviewed `git log` to confirm previous minor changes were likely committed.
+    *   Reviewed active TODOs from `issues.md`.
+    *   Verified TODO #1 (`ProbabilityDistribution` Sorted Inserts) - Deemed sufficiently covered by existing tests (`TestGetQuantization_OrderPreservation`).
+    *   Verified TODO #5 (`ProbabilityDistribution` Disallow `AddPoint` for `Truth`) - Confirmed already implemented and tested (`TestAddPoint_OutOfRangeValueForTruth`).
+    *   Updated `issues.md` to mark TODO #5 as completed for this session date.
+
+3.  **Refactoring Long File (`FileGraphStorageProvider.cs`):**
+    *   Identified `ReasoningEngine/GraphFileHandling/FileGraphStorageProvider.cs` as exceeding the 500-line threshold (initially 660 lines).
+    *   **Phase 1: Path Logic Extraction**
+        *   Created `ReasoningEngine/GraphFileHandling/FilePathHelper.cs`.
+        *   Moved path generation methods (`GetNodeFilePath`, `GetEdgeFilePath`, `GetEdgeDirPath`, `GetIndexFilePath`) and the helper `EnsureDirectoryExists` from `FileGraphStorageProvider` to `FilePathHelper` (as static methods).
+        *   Updated `FileGraphStorageProvider` to use the static methods from `FilePathHelper`.
+        *   Updated `ReasoningEngineTests/GraphFileHandingUnitTests.cs` to use `FilePathHelper` instead of the removed provider methods.
+    *   **Phase 2: Edge Index Logic Extraction**
+        *   Checked line count of `FileGraphStorageProvider` (was 597 lines, still > 500).
+        *   Created `ReasoningEngine/GraphFileHandling/EdgeIndexFileHandler.cs`.
+        *   Moved edge index management methods (`UpdateEdgeIndex`, `RemoveEdgeFromIndex`, `GetAllEdgeFiles`, `LoadIndexFile`, `SaveIndexFile`, `SearchDirectoryForEdges`) and the nested `IndexFile` class to `EdgeIndexFileHandler`.
+        *   Made `IndexFile` public within `EdgeIndexFileHandler` to allow access from tests.
+        *   Updated `FileGraphStorageProvider` to instantiate and use `EdgeIndexFileHandler`.
+        *   Removed the moved methods and the nested `IndexFile` class definition from `FileGraphStorageProvider`.
+        *   Updated `ReasoningEngineTests/GraphFileHandingUnitTests.cs` helper method `LoadIndexFile` and variable declarations to use the public `EdgeIndexFileHandler.IndexFile` type.
+
+4.  **Verification:**
+    *   User confirmed that `dotnet build` and `dotnet test` succeeded after the refactoring steps.
+
+## Next Steps (Paused)
+
+*   Commit the refactoring changes (creation of `FilePathHelper.cs`, `EdgeIndexFileHandler.cs`, and modifications to `FileGraphStorageProvider.cs`, `GraphFileHandingUnitTests.cs`).
+*   Continue addressing remaining TODOs from `issues.md` or other tasks as directed.
